+++ changelog.md
@@ -15,9 +15,9 @@ added to config.json in bixby-backend
 switch to backend and database
 read favorites
 started createSheet
-
-- add a search bar to the top, search by place name or city
-- create favorites
+add a search bar to the top, search by place name or city
+create favorites via heart
+- needs to keep my likes in state and the pop-up when needed
 - update favorites
 - delete favorites
 - list mode